+++ src/Tests/Dfe.Complete.CypressTests/cypress/api/taskApiConversions.ts
@@ -112,6 +112,12 @@ interface UpdatePostDecisionActionsTaskRequest {
     laProformaUploaded?: boolean;
 }
 
+interface UpdateTupeConsultationTaskRequest {
+    taskDataId: TaskDataId;
+    projectType: ProjectType;
+    completed?: boolean;
+}
+
 interface UpdateTrustModificationOrderTaskRequest {
     taskDataId: TaskDataId;
     notApplicable?: boolean;
@@ -188,6 +194,10 @@ class TaskApiConversions extends TaskApi {
 
     public updateTrustModificationOrderTask(requestBody: UpdateTrustModificationOrderTaskRequest) {
         return this.taskDataBaseRequest<void>("TrustModificationOrder", requestBody);
+    }
+
+    public updateTupeConsultationTask(requestBody: UpdateTupeConsultationTaskRequest) {
+        return this.taskDataBaseRequest<void>("TupeConsultation", requestBody);
     }
 }